created utils for api references
