+++ src/components/tina-markdown/embedded-elements/api-reference/api-reference.tsx
@@ -11,79 +11,10 @@ import type {
   ResponseViewState,
   SchemaDetails,
 } from "./types";
+import { generateExample } from "./utils";
 
 // Context to share schema definitions across components
 const SchemaContext = createContext<any>({});
-
-// Helper to resolve $ref references
-export const resolveReference = (ref: string, definitions: any): any => {
-  if (!ref || typeof ref !== "string" || !ref.startsWith("#/")) {
-    return null;
-  }
-
-  // Extract the path from the reference
-  const path = ref.substring(2).split("/");
-
-  // Navigate through the definitions object
-  let result = definitions;
-  for (const segment of path) {
-    if (!result || !result[segment]) {
-      return null;
-    }
-    result = result[segment];
-  }
-
-  return result;
-};
-
-// Move example generation logic to a helper function
-const generateExample = (schema: any, definitions: any, depth = 0): any => {
-  if (depth > 3) return "...";
-  if (schema.$ref) {
-    const refSchema = resolveReference(schema.$ref, definitions);
-    if (refSchema) {
-      return generateExample(refSchema, definitions, depth);
-    }
-    return `<${schema.$ref.split("/").pop()}>`;
-  }
-  switch (schema.type) {
-    case "string":
-      return schema.example || schema.default || "string";
-    case "integer":
-    case "number":
-      return schema.example || schema.default || 0;
-    case "boolean":
-      return schema.example || schema.default || false;
-    case "array":
-      if (schema.items) {
-        const itemExample = generateExample(
-          schema.items,
-          definitions,
-          depth + 1
-        );
-        return [itemExample];
-      }
-      return [];
-    case "object":
-      if (schema.properties) {
-        const obj: any = {};
-        for (const [key, prop] of Object.entries(schema.properties)) {
-          obj[key] = generateExample(prop, definitions, depth + 1);
-        }
-        return obj;
-      }
-      return {};
-    default:
-      if (schema.properties) {
-        const obj: any = {};
-        for (const [key, prop] of Object.entries(schema.properties)) {
-          obj[key] = generateExample(prop, definitions, depth + 1);
-        }
-        return obj;
-      }
-      return null;
-  }
-};
 
 const RequestBodyDropdown = ({ value, onChange }: RequestBodyDropdownProps) => {
   return (
